+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add limited_to_source field to restrict access for external consultants
 - Add privilege field to consultant to implement access filtering
 - Update search, allowing user to search old candidate but not the detail
+- Add upload a file on interview
 
 # v1.16.0 (2022-11-14)
 - More refined factories